[FEAT]: ACPI Platform info parsing Parsed platform info and HPET table, currently only uses HPET table to create a timer.

+++ kernel/src/boot/arch/memory_map.rs
@@ -72,7 +72,7 @@ pub struct BootstrapMemoryMap {
 
 impl BootstrapMemoryMap {
     pub fn new(entries: &[MemoryMapEntry]) -> Self {
-        let size = entries.len().min(Self::SIZE as usize);
+        let size = entries.len().min(Self::SIZE);
         let mut memory_map = Self {
             size: size as u64,
             entries: [MemoryMapEntry::default(); Self::SIZE],
